Upgrade selenium to 3.141.59 and appium java client to 7.0.0
This upgrade try to migrate from old apache HttpClient to OkHttpClient but not all has been migrated successfully (proxy config with credentials for remote driver).
The new version of selenium remove support of HtmlUnit and PhantomJS browser because they has been replaced with Chrome and Firefox Headless which are much better and popular.
Add a TODO flag to migrate the remain apache HttpClient to OkHttpClient also which is lightweight and better.